+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.2 (unreleased)
+## 0.1.2 (2020-04-29)
 
 - Added CIFAR10, CIFAR100, FashionMNIST, and KMNIST datasets
 - Added ResNet18 model